feat: add menu item validation and integrate it into create and update functions in back end
